refactor(app): move useLEDLights
