+++ app/src/organisms/Devices/RobotSettings/AdvancedTab/EnableStatusLight.tsx
@@ -12,7 +12,7 @@ import {
   LegacyStyledText,
 } from '@opentrons/components'
 import { ToggleButton } from '/app/atoms/buttons'
-import { useLEDLights } from '../../hooks'
+import { useLEDLights } from '/app/resources/robot-settings'
 
 interface EnableStatusLightProps {
   robotName: string
