+++ app/src/organisms/Devices/hooks/index.ts
@@ -4,7 +4,6 @@ export * from './useCalibrationTaskList'
 export * from './useIsRobotBusy'
 export * from './useIsRobotViewable'
 export * from './useLights'
-export * from './useLEDLights'
 export * from './useLPCSuccessToast'
 export * from './usePipetteOffsetCalibrations'
 export * from './usePipetteOffsetCalibration'
