+++ app/src/pages/ODD/RobotSettingsDashboard/__tests__/RobotSettingsDashboard.test.tsx
@@ -21,8 +21,8 @@ import {
   UpdateChannel,
 } from '/app/organisms/ODD/RobotSettingsDashboard'
 import { getRobotUpdateAvailable } from '/app/redux/robot-update'
-import { useNetworkConnection } from '/app/resources/networking/hooks/useNetworkConnection'
-import { useLEDLights } from '/app/organisms/Devices/hooks'
+import { useNetworkConnection } from '/app/resources/networking'
+import { useLEDLights } from '/app/resources/robot-settings'
 
 import { RobotSettingsDashboard } from '../'
 
@@ -30,7 +30,8 @@ vi.mock('/app/redux/discovery')
 vi.mock('/app/redux/robot-update')
 vi.mock('/app/redux/config')
 vi.mock('/app/redux/robot-settings')
-vi.mock('/app/resources/networking/hooks/useNetworkConnection')
+vi.mock('/app/resources/networking')
+vi.mock('/app/resources/robot-settings')
 vi.mock('/app/organisms/Navigation')
 vi.mock('/app/organisms/ODD/RobotSettingsDashboard/TouchScreenSleep')
 vi.mock('/app/organisms/ODD/RobotSettingsDashboard/NetworkSettings')
@@ -38,7 +39,6 @@ vi.mock('/app/organisms/ODD/RobotSettingsDashboard/DeviceReset')
 vi.mock('/app/organisms/ODD/RobotSettingsDashboard/RobotSystemVersion')
 vi.mock('/app/organisms/ODD/RobotSettingsDashboard/TouchscreenBrightness')
 vi.mock('/app/organisms/ODD/RobotSettingsDashboard/UpdateChannel')
-vi.mock('/app/organisms/Devices/hooks')
 vi.mock('/app/organisms/ODD/RobotSettingsDashboard/Privacy')
 
 const mockToggleLights = vi.fn()
