+++ app/src/resources/networking/hooks/index.ts
@@ -1,2 +1,3 @@
-export { useCanDisconnect } from './useCanDisconnect'
-export { useWifiList } from './useWifiList'
+export * from './useCanDisconnect'
+export * from './useWifiList'
+export * from './useNetworkConnection'
